chore: rename globalInfo module
