+++ src/stateAggregator/accessors/aliasAccessor.ts
