+++ src/stateAggregator/accessors/tokenAccessor.ts
