+++ src/stateAggregator/globalInfoConfig.ts
