+++ src/stateAggregator/index.ts
