+++ src/stateAggregator/stateAggregator.ts
