+++ src/stateAggregator/types.ts
